\name{listDatabases}
\docType{methods}

\alias{generic,listDatabases}
\alias{listDatabases,LRBaseDb-method}
\alias{listDatabases}

\title{A function to return the scientific name of package}
\description{This function returns the scientific name of species used in the package.}

\usage{listDatabases(x)}

\arguments{
\item{x}{\code{LRBaseDb} object such as \code{LRBase.Mmu.eg.db}}
}
\author{Koki Tsuyuzaki}
\examples{
# library("LRBase.Mmu.eg.db")
# listDatabases(LRBase.Mmu.eg.db)
}