\name{species}
\docType{methods}

\alias{generic,species}
\alias{species,LRBaseDb-method}
\alias{species}

\title{A function to return the name of organism}
\description{This function returns the name of organism.}

\usage{species(object)}

\arguments{
\item{object}{\code{LRBaseDb} object such as \code{LRBase.Mmu.eg.db}}
}
\value{
The common name of organisms
}
\author{Koki Tsuyuzaki}
\examples{
showMethods("species")
}