\name{nomenclature}
\docType{methods}

\alias{generic,nomenclature}
\alias{nomenclature,LRBaseDb-method}
\alias{nomenclature}

\title{A function to return the scientific name}
\description{This function returns the scientific name of
species used in the package.}

\usage{nomenclature(x)}

\arguments{
\item{x}{\code{LRBaseDb} object such as \code{LRBase.Mmu.eg.db}}
}
\value{
The scientific name of organisms
}
\author{Koki Tsuyuzaki}
\examples{
showMethods("nomenclature")
}