\name{packageName}
\docType{methods}

\alias{generic,packageName}
\alias{packageName,LRBaseDb-method}
\alias{packageName}

\title{A function to return the name of package}
\description{This function returns the name of package}

\usage{packageName(x)}

\arguments{
\item{x}{\code{LRBaseDb} object such as \code{LRBase.Mmu.eg.db}}
}
\author{Koki Tsuyuzaki}
\examples{
# library("LRBase.Mmu.eg.db")
# packageName(LRBase.Mmu.eg.db)
}